Replace GuardClauses with ErrorOr
Removes throwing unnecessary exception.

+++ src/SourceName.Application/Commands/DeleteExample.cs
@@ -1,14 +1,17 @@
 ﻿using Ardalis.GuardClauses;
+using ErrorOr;
 using MediatR;
+
+using SourceName.Application.Common.Errors;
 using SourceName.Application.Contracts;
 
 namespace SourceName.Application.Commands;
 
 public static class DeleteExample
 {
-    public record Command(string Id) : IRequest;
+    public record Command(string Id) : IRequest<ErrorOr<Deleted>>;
 
-    public class Handler : IRequestHandler<Command>
+    public class Handler : IRequestHandler<Command, ErrorOr<Deleted>>
     {
         private readonly IRepository _repository;
 
@@ -17,11 +20,17 @@ public static class DeleteExample
             _repository = repository;
         }
 
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<ErrorOr<Deleted>> Handle(Command request, CancellationToken cancellationToken)
         {
             var entity = await _repository.Get(request.Id);
-            Guard.Against.Null(entity);
+
+            if (entity is null)
+            {
+                return Errors.Entity.NotFound;
+            }
+
             await _repository.Delete(entity);
+            return Result.Deleted;
         }
     }
 }
